fix: resolve final Greptile review comments
**Issue 1: Internal review artifact committed to repository**
- Removed CODE_REVIEW_FIXES_SUMMARY.md (auto-generated review tracking file)
- This file was internal housekeeping, not user-facing documentation

**Issue 2: Multiple type declarations in one file**
- Split ReplayMetadata.cs into three separate files following GenHub convention
- Created PlayerInfo.cs for PlayerInfo class
- Created PlayerType.cs for PlayerType enum
- Maintains one-type-per-file consistency with rest of GenHub.Core/Models

**Issue 3: Fragile encoding comparison via reference equality**
- Changed from `encoding == Encoding.Unicode` to `encoding.CodePage == Encoding.Unicode.CodePage`
- Prevents silent corruption if future callers pass new UnicodeEncoding() or Encoding.GetEncoding("utf-16")
- More robust encoding detection for charSize calculation

All Greptile comments now fully resolved!
