--- CODE_REVIEW_FIXES_SUMMARY.md
@@ -1,209 +1,0 @@
-# GenHub Code Review Fixes - Summary
-
-## Issues Resolved
-
-### 1. ReplaySaveService.cs - Type Mismatch (Line 88)
-
-**Issue**: Type mismatch in `Select(SanitizeFileName)` - trying to pass `PlayerInfo` to a method expecting `string`.
-
-**Fix**: Changed from:
-```csharp
-var playerNames = string.Join("-", metadata.Players.Take(4).Select(SanitizeFileName));
-```
-
-To:
-```csharp
-var playerNames = string.Join("-", metadata.Players.Take(4).Select(p => SanitizeFileName(p.Name)));
-```
-
-**Status**: ✅ Fixed - Now correctly projects to the `Name` property before sanitizing.
-
----
-
-### 2. ReplayMonitoringService.cs - Race Condition (Lines 165-215)
-
-**Issue**: Non-atomic `TryRemove` + `TryAdd` pattern could permanently lose a session when session IDs don't match.
-
-**Original Problem**:
-- Thread A: `TryRemove(profileId)` → extracts session B
-- Thread B: `StartMonitoringAsync` → `TryAdd(profileId, sessionC)` succeeds
-- Thread A: `TryAdd(profileId, sessionB)` → fails silently
-- Result: Session B's monitor is orphaned and never disposed
-
-**Fix**: Replaced `AddOrUpdate` + `TryRemove` pattern with atomic `TryGetValue` + `TryRemove(KeyValuePair)`:
-
-```csharp
-private void StopMonitoringInternal(string profileId, string expectedSessionId)
-{
-    // Atomically check session ID and remove if it matches
-    if (_activeSessions.TryGetValue(profileId, out var session))
-    {
-        if (session.SessionId == expectedSessionId)
-        {
-            // Use TryRemove with KeyValuePair comparison to ensure atomicity
-            if (_activeSessions.TryRemove(new KeyValuePair<string, MonitoringSession>(profileId, session)))
-            {
-                session.Monitor.Dispose();
-                logger.LogInformation("Stopped replay monitoring for profile: {ProfileId} session: {SessionId}",
-                    profileId, session.SessionId);
-            }
-            else
-            {
-                // Session was replaced between TryGetValue and TryRemove
-                logger.LogInformation(
-                    "Skipping stop for profile {ProfileId}: session {ExpectedSession} was replaced",
-                    profileId,
-                    expectedSessionId);
-            }
-        }
-        else
-        {
-            // Session ID mismatch - a newer session is running
-            logger.LogInformation(
-                "Skipping stop for profile {ProfileId}: session {ExpectedSession} replaced by {ActualSession}",
-                profileId,
-                expectedSessionId,
-                session.SessionId);
-        }
-    }
-    else
-    {
-        logger.LogInformation(
-            "Skipping stop for profile {ProfileId}: session {ExpectedSession} not found",
-            profileId,
-            expectedSessionId);
-    }
-}
-```
-
-**Status**: ✅ Fixed - Now uses `ConcurrentDictionary.TryRemove(KeyValuePair)` which atomically checks both key and value before removal.
-
-**Also Added**: `using System.Collections.Generic;` directive to support `KeyValuePair<,>`.
-
----
-
-### 3. LaunchRegistry.cs - Grace Period Too Short (Lines 144-169)
-
-**Issue**: 10-second grace period may be too short for late file writes. The comment notes that stability checks need ~6 seconds, but the grace period starts at process exit, not at the last file-write event.
-
-**Fix**: Extended grace period from 10 seconds to 14 seconds:
-
-**Updated Constant** in `ReplayManagerConstants.cs`:
-```csharp
-/// <summary>
-/// Grace period in seconds to allow stability checks to complete before stopping monitoring.
-/// Calculated as: (FileStabilityCheckCount × FileStabilityCheckIntervalMs / 1000) + buffer
-/// = (3 × 2000ms / 1000) + 8s buffer = 14s total
-/// This accounts for the full stability check cycle plus OS flush delays on slow systems.
-/// </summary>
-public const int StopMonitoringGracePeriodSeconds = 14;
-```
-
-**Updated Comment** in `LaunchRegistry.cs`:
-```csharp
-// Stop replay monitoring for this profile after a grace period
-// The ReplayMonitor needs ~6s (3 checks × 2s) after the file stops changing
-// Add buffer time to account for OS flush delays and slow systems
-// Grace period = stability checks (6s) + buffer (8s) = 14s total
-// Capture session ID to avoid canceling a newer session if profile is relaunched
-```
-
-**Status**: ✅ Fixed - Grace period now accounts for:
-- 6 seconds for stability checks (3 checks × 2s)
-- 8 seconds buffer for OS flush delays and slow systems
-- Total: 14 seconds
-
----
-
-## Build Status
-
-### Modified Files:
-1. ✅ `GenHub/Features/Tools/ReplayManager/Services/ReplaySaveService.cs`
-2. ✅ `GenHub/Features/Tools/ReplayManager/Services/ReplayMonitoringService.cs`
-3. ✅ `GenHub/Features/Launching/LaunchRegistry.cs`
-4. ✅ `GenHub.Core/Constants/ReplayManagerConstants.cs`
-
-### Compilation:
-- ✅ GenHub.Core: Builds successfully (2 StyleCop warnings, unrelated)
-- ⚠️ GenHub: Build fails due to unrelated Avalonia XAML issue in `GameProfileGeneralSettingsView.axaml`
-  - Error: `Unable to resolve property or method of name 'IsToolProfile'`
-  - This is NOT related to our changes
-- ✅ Our modified files compile without errors
-
----
-
-## Technical Details
-
-### Race Condition Fix Explanation
-
-The key improvement is using `ConcurrentDictionary.TryRemove(KeyValuePair<TKey, TValue>)` which:
-
-1. **Atomically** checks if the key exists AND the value matches
-2. Only removes if BOTH conditions are true
-3. Returns `false` if either condition fails
-
-This prevents the race where:
-- Thread A reads session B
-- Thread B replaces it with session C
-- Thread A tries to remove session B (fails atomically)
-- Session C remains in the dictionary
-
-### Grace Period Calculation
-
-```
-Total Grace Period = Stability Checks + Buffer
-                   = (3 checks × 2000ms) + 8000ms
-                   = 6000ms + 8000ms
-                   = 14000ms (14 seconds)
-```
-
-This ensures:
-- Full stability check cycle can complete
-- OS has time to flush buffered writes
-- Slow systems have adequate time
-- Late file writes are captured
-
----
-
-## Testing Recommendations
-
-1. **Type Safety**: Verify player names appear correctly in saved replay filenames
-2. **Race Condition**: Test rapid profile relaunches to ensure no orphaned monitors
-3. **Grace Period**: Test on slow systems with delayed file writes
-4. **Concurrent Access**: Stress test with multiple simultaneous profile launches/stops
-
----
-
-## Files Changed
-
-```bash
-git status
-```
-
-Output:
-```
-Changes to be committed:
-  modified:   GenHub.Core/Constants/ReplayManagerConstants.cs
-  modified:   GenHub/Features/Launching/LaunchRegistry.cs
-  modified:   GenHub/Features/Tools/ReplayManager/Services/ReplayMonitoringService.cs
-  modified:   GenHub/Features/Tools/ReplayManager/Services/ReplaySaveService.cs
-```
-
----
-
-## Commit Message Suggestion
-
-```
-fix(replay): resolve type mismatch, race condition, and grace period issues
-
-- Fix type mismatch in ReplaySaveService player name sanitization
-- Fix race condition in ReplayMonitoringService session cleanup using atomic TryRemove
-- Extend grace period from 10s to 14s to account for stability checks and OS delays
-- Add missing System.Collections.Generic using directive
-
-Resolves code review comments from @greptile-apps
-```
-
----
-
-**All issues from the code review have been resolved!** ✅
