+++ GenHub/GenHub.Core/Models/Tools/ReplayManager/PlayerType.cs
@@ -1,0 +1,22 @@
+namespace GenHub.Core.Models.Tools.ReplayManager;
+
+/// <summary>
+/// Player type enumeration.
+/// </summary>
+public enum PlayerType
+{
+    /// <summary>
+    /// Human player.
+    /// </summary>
+    Human,
+
+    /// <summary>
+    /// AI/Computer player.
+    /// </summary>
+    Computer,
+
+    /// <summary>
+    /// Observer/spectator.
+    /// </summary>
+    Observer
+}
